taquito: compute estimated fees correctly in case of reveal
taquito ignores implicit reveal during estimation so it has to be done by
hand. `getManagerKey` returns null if pkh has not been reveal. We use this
function to check, and add the default fee for reveal to total fee.

A new field is also added if we want to display a specific line for
reveal in the simulation result in the future.